+++ wpfrank-companion/inc/homerix/customizer/page-sections/contact-us-page-section.php
@@ -232,13 +232,13 @@ function homerix_contact_us_page_controls() {
 					'card_title'       => __( 'Email Us', 'homerix' ),
 					'card_description' => __( "Send us a message and we'll respond within 24 hours.", 'homerix' ),
 					'item1_icon'       => 'fas fa-inbox',
-					'item1_value'      => 'info@Homerixpro.com',
+					'item1_value'      => 'info@example.com',
 					'item1_link_type'  => 'mailto',
 					'item2_icon'       => 'fas fa-headset',
-					'item2_value'      => 'support@Homerixpro.com',
+					'item2_value'      => 'support@example.com',
 					'item2_link_type'  => 'mailto',
 					'item3_icon'       => 'fas fa-briefcase',
-					'item3_value'      => 'careers@Homerixpro.com',
+					'item3_value'      => 'careers@example.com',
 					'item3_link_type'  => 'mailto',
 				),
 				// Visit Us Card.
@@ -744,3 +744,4 @@ function homerix_contact_us_page_controls() {
 // Initialize sections and controls.
 add_action( 'init', 'homerix_contact_us_page_section' );
 add_action( 'init', 'homerix_contact_us_page_controls' );
+