+++ wpfrank-companion/inc/homerix/page-templates/page-terms-of-service.php
@@ -78,7 +78,7 @@ $sections = get_theme_mod(
 // Contact Information.
 $contact_title           = get_theme_mod( 'terms_contact_title', __( 'Contact Information', 'homerix' ) );
 $contact_subtitle        = get_theme_mod( 'terms_contact_subtitle', __( 'If you have any questions about these Terms, please contact us:', 'homerix' ) );
-$contact_email           = get_theme_mod( 'terms_contact_email', 'legal@Homerixpro.com' );
+$contact_email           = get_theme_mod( 'terms_contact_email', 'legal@example.com' );
 $contact_phone_enabled   = get_theme_mod( 'terms_contact_phone_enabled', true );
 $contact_phone           = get_theme_mod( 'terms_contact_phone', '(555) 123-4567' );
 $contact_address_enabled = get_theme_mod( 'terms_contact_address_enabled', true );
@@ -197,3 +197,4 @@ $additional_content      = HOMERIX_IS_PRO() ? get_theme_mod( 'terms_additional_c
 
 <?php
 get_footer();
+